fix: avoid NPE in WalletClientSystem by using a binding (#23)
`preBegin()`, despite its description, may be called only after the system already received events to its event handler. In that case, the `walletHud` was still null and causing an NPE in `onUpdateWallet`.

By using a binding this NPE is avoided. The binding value can be updated from the event handler even before it was hooked up with the HUD element.